requirements.txt
alembic==1.13.1
blinker==1.8.0
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.0
Flask-HTTPAuth==4.8.0
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.3
Mako==1.3.3
MarkupSafe==2.1.5
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.1
pytz==2024.1
SQLAlchemy==2.0.29
typing_extensions==4.11.0
Werkzeug==3.0.2
